build kubelet flags directly
